refactor : refactor code related to websocket on client side
